requirements.txt
# QueryNest – FAISS-first RAG CLI (Python >= 3.10,<3.13)

# LangChain core
langchain==0.3.27
langchain-core==0.3.80
langchain-community==0.3.31
langchain-text-splitters==0.3.11

# Google Gemini integration
langchain-google-genai==2.0.10
google-generativeai==0.8.5

# Environment configuration
python-dotenv==1.2.1

# Vector store (FAISS – local)
faiss-cpu==1.8.0.post1

# Tokenization and validation
tiktoken==0.12.0
pydantic==2.12.5

# YouTube transcript loader
youtube-transcript-api==0.6.2

# PDF loader
pypdf==6.4.0

# Web content loading and cleaning
requests==2.32.5
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.2.2

# for progress bars
tqdm
